app/requirements.txt
fastapi
uvicorn
databricks-sdk
psycopg2-binary
requests
